added some more photos + changed how new token gen is done
