+++ src/api/syncData.js
@@ -1,15 +1,9 @@
 const fs = require('fs');
-const readline = require('readline');
-const { google } = require('googleapis');
 const batteryDataModel = require('../schemas/batterySchema.js');
-const { scopes, sheetURL } = require('../../config/constants/googleAPIConstants.json');
-
-// If modifying scopes, delete token.json.
-// The file token.json stores the user's access and refresh tokens, and is
-// created automatically when the authorization flow completes for the first
-// time.
+const { sheetURL } = require('../../config/constants/googleAPIConstants.json');
 const TOKEN_PATH = './config/auth/token.json';
 const CREDENTIALS_PATH = './config/auth/credentials.json'
+const { google } = require('googleapis');
 
 async function checkDb() {
     const res = await batteryDataModel.find({ updated: false })
@@ -92,43 +86,14 @@ async function authorize(credentials, callback, values, document, row) {
 
     // Check if we have previously stored a token.
     fs.readFile(TOKEN_PATH, (err, token) => {
-        if (err) return getNewToken(oAuth2Client, callback, values, document, row);
+        if (err) {
+            console.error('ERROR: No Token. Please run getToken.js')
+            return process.kill()
+        }
         oAuth2Client.setCredentials(JSON.parse(token));
         callback(oAuth2Client, values, document, row);
     });
 }
-
-/**
- * Get and store new token after prompting for user authorization, and then
- * execute the given callback with the authorized OAuth2 client.
- * @param {google.auth.OAuth2} oAuth2Client The OAuth2 client to get token for.
- * @param {getEventsCallback} callback The callback for the authorized client.
- */
-function getNewToken(oAuth2Client, callback, values, document, row) {
-    const authUrl = oAuth2Client.generateAuthUrl({
-        access_type: 'offline',
-        scope: scopes,
-    });
-    console.log('Authorize this app by visiting this url:', authUrl);
-    const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-    });
-    rl.question('Enter the code from that page here: ', (code) => {
-        rl.close();
-        oAuth2Client.getToken(code, (err, token) => {
-            if (err) return console.error('Error while trying to retrieve access token', err);
-            oAuth2Client.setCredentials(token);
-            // Store the token to disk for later program executions
-            fs.writeFile(TOKEN_PATH, JSON.stringify(token), (err) => {
-                if (err) return console.error(err);
-                console.log('Token stored to', TOKEN_PATH);
-            });
-            callback(oAuth2Client, values, document, row);
-        });
-    });
-}
-
 /**
  * @param {google.auth.OAuth2} auth The authenticated Google OAuth client.
  */
