+++ src/index.js
@@ -2,7 +2,7 @@ const runSync = require('./api/runSync.js');
 const connect = require('./database/connect.js');
 try {
     connect()
-} catch(err) { if (err) console.error(err); };
+} catch(err) { if (err) console.log(err); };
 
 try {
     runSync();
