+++ src/tools/getNewToken.js
@@ -1,0 +1,5 @@
+getNewToken = require('../api/getToken.js');
+
+try {
+    getNewToken();
+} catch(err) { console.log(err) };
